+++ lib/variance.dart
@@ -32,6 +32,6 @@ part 'src/4337/userop.dart';
 part 'src/4337/wallet.dart';
 part 'src/common/factory.dart';
 part 'src/common/plugins.dart';
-part 'src/signers/credential_signer.dart';
+part 'src/signers/private_key_signer.dart';
 part 'src/signers/hd_wallet_signer.dart';
 part 'src/signers/passkey_signer.dart';