+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.0.4
+
+* Added a Secure Storage Repository for saving credentials to encrypted android shared-preference and iOS keychain
+
+* Biometric Auth layer over the secure storage repository
+
+* Deprecate the ChainbaseAPI helpers for ENS resolution to use the new ensdart_v3
+
+* Bug Fixes
+
+## 0.0.3
+
+* Temporarily removed support for internally handling local authentication and credential storage in the hd signer
+
 ## 0.0.1
 
 * Create SmartWallet class using simple accounts